% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd.R
\name{cmd_tick}
\alias{cmd_tick}
\title{Tell the runtime to send a tick message after \code{interval} seconds.}
\usage{
cmd_tick(interval)
}
\arguments{
\item{interval}{Seconds until the tick fires.}
}
\description{
Tell the runtime to send a tick message after \code{interval} seconds.
}
