% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{is_key}
\alias{is_key}
\title{Is a message a (specific) keypress?}
\usage{
is_key(msg, key = NULL)
}
\arguments{
\item{msg}{A message.}

\item{key}{Optional key name to match (e.g. \code{"q"}, \code{"up"}).}
}
\description{
Is a message a (specific) keypress?
}
