% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmd.R
\name{with_cmd}
\alias{with_cmd}
\title{Model with a command}
\usage{
with_cmd(model, cmd, ...)
}
\arguments{
\item{model}{A model.}

\item{cmd}{A command.}

\item{...}{Passed to methods.}
}
\description{
A model with a command and a command to execute.
}
